+++ packages/@aws-cdk-testing/cli-integ/tests/cli-integ-tests/cdk-cdk-ls---show-dependencies---json.integtest.ts
@@ -81,3 +81,15 @@ integTest(
   }),
 );
 
+/**
+ * Type to store stack dependencies recursively
+ */
+type DependencyDetails = {
+  id: string;
+  dependencies: DependencyDetails[];
+};
+
+type StackDetails = {
+  id: string;
+  dependencies: DependencyDetails[];
+};
