+++ packages/@aws-cdk-testing/cli-integ/tests/cli-integ-tests/cdk-cdk-ls.integtest.ts
@@ -34,17 +34,3 @@ integTest(
     }
   }),
 );
-
-/**
- * Type to store stack dependencies recursively
- */
-type DependencyDetails = {
-  id: string;
-  dependencies: DependencyDetails[];
-};
-
-type StackDetails = {
-  id: string;
-  dependencies: DependencyDetails[];
-};
-
