+++ packages/@aws-cdk-testing/cli-integ/tests/cli-integ-tests/cdk-generating-and-loading-assembly.integtest.ts
@@ -1,7 +1,7 @@
 import { promises as fs } from 'fs';
 import * as os from 'os';
 import * as path from 'path';
-import { integTest, shell, withDefaultFixture } from '../../lib';
+import { integTest, withDefaultFixture } from '../../lib';
 
 jest.setTimeout(2 * 60 * 60_000); // Includes the time to acquire locks, worst-case single-threaded runtime
