+++ packages/@aws-cdk-testing/cli-integ/tests/cli-integ-tests/cdk-requests-go-through-a-proxy-when-configured.integtest.ts
@@ -2,6 +2,7 @@ import { promises as fs } from 'fs';
 import * as os from 'os';
 import * as path from 'path';
 import { integTest, withDefaultFixture } from '../../lib';
+import { awsActionsFromRequests, startProxyServer } from '../../lib/proxy';
 
 jest.setTimeout(2 * 60 * 60_000); // Includes the time to acquire locks, worst-case single-threaded runtime
