+++ packages/@aws-cdk-testing/cli-integ/tests/cli-integ-tests/cdk-security-related-changes-without-a-cli-are-expected-to-fail.integtest.ts
@@ -1,5 +1,5 @@
 import { DescribeStacksCommand } from '@aws-sdk/client-cloudformation';
-import { integTest, shell, withDefaultFixture } from '../../lib';
+import { integTest, withDefaultFixture } from '../../lib';
 
 jest.setTimeout(2 * 60 * 60_000); // Includes the time to acquire locks, worst-case single-threaded runtime
