+++ packages/@aws-cdk-testing/cli-integ/tests/cli-integ-tests/cdk-templates-on-disk-contain-metadata-resource.integtest.ts
@@ -1,4 +1,4 @@
-import { integTest, shell, withDefaultFixture } from '../../lib';
+import { integTest, withDefaultFixture } from '../../lib';
 
 jest.setTimeout(2 * 60 * 60_000); // Includes the time to acquire locks, worst-case single-threaded runtime
